requirements.txt
Pillow~=9.4.0
PySide6~=6.4.0.1
matplotlib~=3.6.3
numpy~=1.24.2
pandas~=1.5.1